// types/form.ts
export interface UserFormData {
  name: string
  age: string
  gender: string
  height: string
  weight: string
  sleepHours: string
  waterIntake: string
  fitnessGoal: string
  fitnessLevel: string
  workoutLocation: string
  stressLevel: string
  dietaryPreference: string
  medicalHistory: string
}